requirements.txt
matplotlib~=3.2.2
numpyencoder~=0.3.0
numpy~=1.23.5
pandas~=1.4.4
Pillow~=9.3.0
scikit-learn~=1.3.0
umap-learn~=0.5.4
